requirements.txt
# RAG Teaching Assistant — Python Dependencies
# Compatible with Python 3.9+
# Install with: pip install -r requirements.txt

# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# YouTube download
yt-dlp>=2024.1.1

# Speech-to-text (Whisper)
openai-whisper>=20240930
ffmpeg-python>=0.2.0

# LLM + RAG
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4

# Reranker
sentence-transformers>=2.7.0

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0.0
httpx>=0.27.0
